feat(js-core): extract BuckarooView + IModel for injectable transport (#759)
Split the rendering core out of BuckarooServerView into a new
transport-agnostic BuckarooView. The new component takes an already-
connected IModel plus the initial_state payload, so hosts that need to
keep WebSockets out of the renderer (Tauri / Electron / Wails relaying
through IPC) can mount BuckarooView with their own adapter-backed model.

- New IModel interface (send/get/set/save_changes/on/off) — the minimal
  surface BuckarooView consumes. WebSocketModel is structurally
  assignable to it.
- New BuckarooView component — owns state hooks, model change-event
  wiring, row-cache construction, and widget dispatch by mode.
- BuckarooServerView keeps its public API byte-for-byte. It still
  constructs WebSocketModel, awaits initial_state, runs
  preResolveDFDataDict, and now delegates rendering to BuckarooView.
- Export BuckarooView, IModel, and the prop/type aliases from the
  package entry point.
